pyproject.toml
[project]
name = "coleman"
version = "2.3.0"
description = "Multi-Armed Bandit based test case prioritization for continuous integration environments"
license = "MIT"
authors = [{ name = "Jackson Antonio do Prado Lima", email = "jacksonpradolima@gmail.com" }]
requires-python = ">=3.14"
readme = "README.md"
dependencies = [
    "duckdb",
    "numpy",
    "polars",
    "pyarrow",
    "scipy",
    "scikit-posthocs",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "pyright",
    "ty",
    "pre-commit",
    "interrogate",
    "numpydoc",
    "pytest",
    "pytest-benchmark",
    "pytest-cov",
    "pytest-mock",
    "pytest-profiling",
    "pytest-xdist",
    "radon>=6.0.1",
    "xenon>=0.9.3",
    "scalene>=2.2.1",
    "py-spy>=0.4.1",
    "codecarbon>=3.2.6",
    "wily>=1.12.2",
    "pyrapl>=0.2.3.1",
]
docs = [
    "zensical",
    "mkdocstrings[python]",
    "mkdocs-jupyter",
]
release = [
    "python-semantic-release",
]
notebook = [
    "marimo",
    "matplotlib",
    "pandas",
    "seaborn",
]
telemetry = [
    "opentelemetry-api",
    "opentelemetry-sdk",
    "opentelemetry-exporter-otlp-proto-http",
]
clickhouse = [
    "clickhouse-connect",
]

[project.scripts]
coleman = "coleman.cli:main"

[project.urls]
homepage = "https://github.com/jacksonpradolima/coleman"
repository = "https://github.com/jacksonpradolima/coleman"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["coleman", "coleman4hcs"]

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "N", "D"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["D100", "D101", "D102", "D103", "D104", "D107", "D200", "D205", "D400", "D401"]
"notebooks/**" = ["D", "E501", "F821", "F841", "B023", "B905", "N", "UP", "SIM", "W"]
"coleman/exceptions.py" = ["N818"]
"examples/**" = ["D", "E501"]
"scripts/**" = ["D"]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.pyright]
pythonVersion = "3.14"
typeCheckingMode = "basic"

[tool.pytest.ini_options]
addopts = [
    "--durations=25",
    "--durations-min=0.05",
    "--benchmark-min-rounds=1",
    "--benchmark-max-time=0.05",
    "--benchmark-sort=mean",
]
markers = [
    "performance: Mark a test as a performance benchmark.",
]

[tool.interrogate]
ignore-init-module = true
fail-under = 80

[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
branch = "main"
commit_message = "chore(release): {version}"
build_command = "uv build"
